+++ ai_ref_kits/agentic_llm_rag/config/paint_concierge_personality.yaml
