+++ ai_ref_kits/agentic_llm_rag/data/test_painting_llm_rag.pdf
